pyproject.toml
[project]
name = "my-voice-tts"
version = "0.1.0"
description = "Local voice cloning TTS - read blog posts in your voice"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "chatterbox-tts>=0.1.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "rich>=13.0.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "pydub>=0.25.1",
]

[project.scripts]
my-voice-tts = "my_voice_tts.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/my_voice_tts"]
